+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-08 16:35:01
+
+### Changed Files
+- `src/features/policies/PolicyForm.tsx`
+- `src/features/policies/components/PolicyDialog.tsx`
+
+### Statistics
+```
+ 2 files changed, 54 insertions(+), 17 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-08 16:26:35
 
 ### Changed Files